# apps/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ご近所サポートちょこっと API設計書
  version: 1.0.0
  description: |
    高齢者 × ボランティア マッチングアプリ「ちょこっと」のAPI仕様書です。  
    各エンドポイントの概要・入力・出力・ステータスコードを示します。
servers:
  - url: http://localhost:3001/api
    description: 開発環境
...
paths:
...
  /supporter/orders:
    get:
      tags:
        - Supporter
      summary: サポーターが引き受けた依頼一覧を取得
      description: |
        ログイン中のサポーター自身が引き受けた依頼一覧を返します。  
        依頼者情報を含みます。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 引受一覧取得成功（空の場合は空配列を返します）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupporterOrderSummary"
        "403":
          description: サポーターのみ閲覧できます。
        "500":
          description: サーバーエラー
  /supporter/orders/{requestId}:
    post:
      tags:
        - Supporter
      summary: 依頼を引き受ける
      description: |
        指定した requestId の依頼を "matched" に更新し、  
        supporterId を設定した Orderデータを作成します。  
        依頼の status が open の場合のみ引受可能です。
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
            example: 101
      responses:
        "201":
          description: 引受登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupporterOrderCreated"
        "400":
          description: この依頼は受付終了です。
        "403":
          description: サポーターのみ引受可能です。
        "404":
          description: 指定された依頼が見つかりません。
        "500":
          description: サーバーエラー
  /orders/{requestId}:
    patch:
      tags:
        - Orders
      summary: 依頼のステータスを更新する  
      description: |
        指定された requestId の依頼に対して、status の更新を行います。    
        実行者の role（user / supporter）によって更新できるステータスが異なります。  
        
        **利用者(user) が更新できるステータス：**  
        - canceled  
        - refusal（サポーター交代要請）

        **サポーター(supporter) が更新できるステータス：**  
        - confirmed（電話連絡済み）  
        - completed  
        - decline（辞退）

        **decline / refusal の場合：**  
        - order.status を decline/refusal に変更  
        - request.status を open に戻す  
        - request.matchedSupporterId を null に戻す
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: integer
            example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [updateStatus]
              properties:
                updateStatus:
                  type: string
                  enum:
                    - confirmed
                    - completed
                    - canceled
                    - decline
                    - refusal
                  example: "confirmed"
      responses:
        "200":
          description: ステータス更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  OrderRequestUpdated:
                    $ref: "#/components/schemas/OrderRequestUpdated"

              examples:
                createdOrder:
                  summary: Ordersの更新
                  value:
                    status: "canceled"
                updatedRequest:
                  summary: Requestの更新
                  value:
                    status: "canceled"

        "400":
          description: |
            バリデーションエラー  
            - リクエストIDは数値を指定してください。  
            - 不正なステータスです。
        "403":
          description: |
            実行権限なし  
            - 依頼者本人のみ操作可能です。  
            - 担当サポーターのみ操作可能です。
        "404":
          description: 指定された依頼が見つかりません。
        "409":
          description: |
            不整合  
            - 引受されていないため、この操作はできません。  
            - 対応する引受データが見つかりません。  
        "500":
          description: サーバーエラー
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SupporterOrderSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, matched, confirmed, completed, canceled, expired]
          example: "matched"
        scheduledDate:
          type: string
          format: date
        scheduledStartTime:
          type: string
        scheduledEndTime:
          type: string
        scheduledDurationMinutes:
          type: integer
          nullable: true
        workLocation1:
          type: string
        workLocation2:
          type: string
        matchedSupporterId:
          type: string
          nullable: true
        centerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: string
            gender:
              type: string
            address1:
              type: string
            ageGroup:
              type: string
              example: "80代"
            bio:
              type: string

    # =====================================================
    # 引受時の返却形式 SupporterOrderCreated
    # =====================================================
    SupporterOrderCreated:
      type: object
      properties:
        message:
          type: string
          example: "Order created successfully"

        updatedRequest:
          type: object
          properties:
            matchedSupporterId:
              type: string
            status:
              type: string
              enum: [open, matched, confirmed, completed, canceled]
              example: "matched"

        createdOrder:
          type: object
          properties:
            requestId:
              type: integer
            supporterId:
              type: string
            status:
              type: string
              enum: [matched, confirmed, completed, canceled, decline, refusal]
              example: "matched"
    
    # =====================================================
    # ステータス更新時の返却形式 OrderRequestUpdated
    # =====================================================
    OrderRequestUpdated:
      type: object
      properties:
        message:
          type: string
          example: "Status updated successfully"

        updatedOrder:
          type: object
          properties:
            requestId:
              type: integer
            supporterId:
              type: string
            status:
              type: string
              enum: [matched, confirmed, completed, canceled, decline, refusal]
              example: "canceled"

        updatedRequest:
          type: object
          properties:
            matchedSupporterId:
              type: string
            status:
              type: string
              enum: [open, matched, confirmed, completed, canceled]
              example: "canceled"

    # =====================================================
    # ErrorResponse
    # =====================================================
...